Sprint2_day2
Draft of shooting function code. Able to create moving projectile images however no collider yet.
