+++ Super Iskolar/Assets/Prefabs/Pencil.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6ffce11bc5166904c8a98d4f934b1236
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
