+++ Super Iskolar/Assets/Prefabs/UPOvalBench.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4227f3d9d75bfbf478d0a356b5ce9e65
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
