+++ Super Iskolar/Assets/Scripts/BulletScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 891693fc8069dd547b063b1d1e92e22e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
