+++ Super Iskolar/Assets/Scripts/InterpretInputControllerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 508fbd897f0c9dc4e95afd66e19d249e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
