--- Super Iskolar/Assets/Scripts/PencilWeapon.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PencilWeapon : MonoBehaviour
-{
-
-    public Transform FirePencil;
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
