--- Super Iskolar/Assets/Scripts/PencilWeapon.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a7361c4b2e80662458145f5494b9785b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
