+++ Super Iskolar/Assets/Scripts/ShootScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7361c4b2e80662458145f5494b9785b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
