+++ Super Iskolar/Logs/shadercompiler-AssetImportWorker0.log
@@ -1,3 +1,3 @@
-Base path: '/Applications/Unity/Hub/Editor/2020.3.29f1/Unity.app/Contents', plugins path '/Applications/Unity/Hub/Editor/2020.3.29f1/PlaybackEngines'
+Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.29f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.29f1/Editor/Data/PlaybackEngines'
 Cmd: initializeCompiler
