+++ Super Iskolar/Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,0 +1,3 @@
+Base path: 'C:/Program Files/Unity/Hub/Editor/2020.3.29f1/Editor/Data', plugins path 'C:/Program Files/Unity/Hub/Editor/2020.3.29f1/Editor/Data/PlaybackEngines'
+Cmd: initializeCompiler
+
