--- Super Iskolar/Logs/shadercompiler-UnityShaderCompiler0.log
@@ -1,3 +1,0 @@
-Base path: '/Applications/Unity/Hub/Editor/2020.3.29f1/Unity.app/Contents', plugins path '/Applications/Unity/Hub/Editor/2020.3.29f1/PlaybackEngines'
-Cmd: initializeCompiler
-
